Refactor widget classes
